plugin: move Expr filter to its own module
In preparation for adding other filters.
